+++ map_arm/package.xml
@@ -11,6 +11,7 @@
   
   <depend>rclcpp</depend>
   <depend>visualization_msgs</depend>
+  <build_depend>filter_points</build_depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
